+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/Checkbox/CheckboxIndeterminatePage.swift
@@ -50,6 +50,7 @@ private struct CheckboxIndeterminateDemo: View {
                                       accessibilityLabel: "app_components_checkbox_hint_a11y".localized(), // No LocalizedStringKey inference in the component
                                       isError: configurationModel.isError)
                 .disabled(!configurationModel.enabled)
+                .accessibilityIdentifier(A11YIdentifiers.componentCheckboxIndeterminate)
             Spacer()
         }
         .padding(.all, theme.spaces.spaceFixedMedium)