+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/Checkbox/CheckboxPage.swift
@@ -50,6 +50,7 @@ private struct CheckboxDemo: View {
                          accessibilityLabel: "app_components_checkbox_hint_a11y".localized(), // No LocalizedStringKey inference in the component
                          isError: configurationModel.isError)
                 .disabled(!configurationModel.enabled)
+                .accessibilityIdentifier(A11YIdentifiers.componentCheckbox)
             Spacer()
         }
         .padding(.all, theme.spaces.spaceFixedMedium)