+++ DesignToolbox/DesignToolbox/Pages/Components/Radio/RadioItem/RadioItemPage.swift
@@ -56,6 +56,7 @@ private struct RadioItemDemo: View {
                       hasDivider: configurationModel.divider)
             .disabled(!configurationModel.enabled)
             .padding(.all, theme.spaces.spaceFixedMedium)
+            .accessibilityIdentifier(A11YIdentifiers.componentRadioItem)
     }
 
     // Need here that system name, a11y managed in component