+++ DesignToolbox/DesignToolbox/Pages/Components/Switch/SwitchItem/SwitchItemPage.swift
@@ -52,6 +52,7 @@ private struct SwitchItemDemo: View {
                        hasDivider: configurationModel.divider)
             .disabled(!configurationModel.enabled)
             .padding(.all, theme.spaces.spaceFixedMedium)
+            .accessibilityIdentifier(A11YIdentifiers.componentSwitchItem)
     }
 
     // Need here that system name, a11y managed in component