+++ DesignToolbox/DesignToolbox/Utils/A11YIdentifiers.swift
@@ -1,0 +1,31 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+/// Just to gather accessiblity identifiers for components of the app for UI automated tests
+enum A11YIdentifiers {
+
+    // MARK: - Components
+
+    static let componentCheckbox = "designSystemToolbox_component_checkbox"
+    static let componentCheckboxItem = "designSystemToolbox_component_checkboxItem"
+    static let componentCheckboxIndeterminate = "designSystemToolbox_component_checkboxIndeterminate"
+    static let componentCheckboxItemIndeterminate = "designSystemToolbox_component_checkboxItemIndeterminate"
+    static let componentRadio = "designSystemToolbox_component_radio"
+    static let componentRadioItem = "designSystemToolbox_component_radioItem"
+    static let componentSwitch = "designSystemToolbox_component_switch"
+    static let componentSwitchItem = "designSystemToolbox_component_switchitem"
+
+    // MARK: - Configuration of components (demo)
+
+    static let enableSwitch = "designSystemToolbox_configuration_switch_enable"
+}